requirements.txt
flask==2.3.3
flask-cors==4.0.0
gtts==2.4.0
pdfplumber==0.10.3
python-docx==0.8.11
ebooklib==0.18
beautifulsoup4==4.12.2
requests==2.31.0
spacy==3.7.2
transformers==4.35.0
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
nltk==3.8.1